+++ CHANGELOG.md
@@ -8,6 +8,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ### Added
 * Added `cargoNextest` for running tests via [cargo-nextest](https://nexte.st/)
+* Added `cargoAudit` for running [cargo-audit](https://crates.io/crates/cargo-audit)
+  with a provided advisory database instance.
 
 ### Changed
 * **Breaking**: the `--workspace` flag is no longer set for all cargo commands